Begin expicitly declaring mix folders
These will differ based on how you have your mixes named.
I am doing it according to how mine are named,
and if you don't like it, that's sad for you.

It's been a long time since I actually made this change, so I can't
remember how I did it, but very likely with some hacky shell script,
so some of the changes may be wrong. But hopefully a step forward.